fix(condo): DOMA-6740 move attach call record button to action bar
